+++ core/forms.py
@@ -11,10 +11,11 @@ from directory_validators.company import no_html
 from captcha.fields import ReCaptchaField
 
 
-from directory_constants.choices import COUNTRY_CHOICES
-
+from directory_constants.choices import COUNTRY_CHOICES, EMPLOYEES, INDUSTRIES
 
 COUNTRIES = BLANK_CHOICE_DASH + COUNTRY_CHOICES
+COMPANY_SIZE = BLANK_CHOICE_DASH + list(EMPLOYEES)
+INDUSTRY_OPTIONS = BLANK_CHOICE_DASH + list(INDUSTRIES)
 
 
 class TariffsCountryForm(forms.Form):
@@ -126,6 +127,59 @@ class CapitalInvestContactForm(GovNotifyActionMixin, forms.Form):
         # `captcha` and `terms_agreed` are not useful to agent
         # as those fields have to be present
         # for the form to be submitted.
+        data = self.cleaned_data.copy()
+        del data['captcha']
+        del data['terms_agreed']
+        return data
+
+
+class BusinessEnvironmentGuideForm(GovNotifyActionMixin, forms.Form):
+    given_name = forms.CharField(label=_('Given name'), required=True)
+    family_name = forms.CharField(label=_('Family name'), required=True)
+    email_address = forms.EmailField(label=_('Email address'), required=True)
+    company_name = forms.CharField(label=_('Company name (Optional)'), required=False)
+    number_of_staff = forms.ChoiceField(
+        label=_('Current number of staff (Optional)'),
+        choices=COMPANY_SIZE,
+        required=False,
+    )
+    industry = forms.ChoiceField(
+        label=_('Industry'),
+        choices=INDUSTRY_OPTIONS,
+        required=True,
+    )
+    country = forms.ChoiceField(
+        label=_('Country'),
+        widget=Select(attrs={'id': 'js-country-select'}),
+        choices=COUNTRIES
+    )
+
+    mostly_interested_in = forms.MultipleChoiceField(
+        label=_('I am mainly interested in:'),
+        help_text=_('Select all that apply'),
+        widget=forms.CheckboxSelectInlineLabelMultiple(),
+        choices=(
+            ('expand', _('Setting up a business in the UK')),
+            ('invest_capital', _('Investing capital in the UK')),
+            ('buy_goods', _('Buying good from the UK')),
+        ),
+        required=True
+    )
+
+    further_information = forms.BooleanField(
+        label=_('I would like to receive further information'),
+        required=False,
+    )
+    terms_agreed = forms.BooleanField(
+        label=_('I confirm the information provided is true and I accept DIT\'s terms and conditions'),
+        required=True,
+    )
+
+    captcha = ReCaptchaField(label='', label_suffix='')
+
+    @property
+    def serialized_data(self):
+        # `captcha` and `terms_agreed` are not useful to the agent so remove them from the submitted data.
         data = self.cleaned_data.copy()
         del data['captcha']
         del data['terms_agreed']