+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - CI-452 - Added new international home page template to use if `is_new_page_ready` is present and True
 - CI-479 - Added video to Why choose UK Page
 - CI-474 - Added feature flagged redirects and temporary look up by paths for when pages move in the cms
+- CI-308 - Added UK Business Environment Form and success page behind a feature flag.
 
 ### Fixed bugs
 - no ticket - change FAS search input broken id